+++ examples/scripts/hello_world.py
@@ -44,8 +44,9 @@ if __name__ == "__main__":
     # provider account creation can be skipped in case you already saved
     # your provider account to the `~/.qiskit/tergiterc` file.
     # See below how that is done.
-    account = AccountInfo(service_name=SERVICE_NAME, url=API_URL, token=API_TOKEN)
-    provider = Tergite.use_provider_account(account)
+    provider = Tergite.use_provider_account(
+        service_name=SERVICE_NAME, url=API_URL, token=API_TOKEN
+    )
     # to save this account to the `~/.qiskit/tergiterc` file, add the `save=True`
     # provider = Tergite.use_provider_account(account, save=True)