+++ tests/integration/test_provider.py
@@ -15,10 +15,9 @@ _PROVIDER_ACCOUNTS = load_json_fixture("provider_accounts.json")
 @pytest.mark.parametrize("account_data", _PROVIDER_ACCOUNTS)
 def test_use_provider_account(account_data, mock_tergiterc):
     """use_provider_account without save option returns the given provider account"""
-    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
-    provider = _tergite.use_provider_account(account)
-    assert provider.account == account
+    provider = _tergite.use_provider_account(**account_data)
+    assert provider.account == AccountInfo(**account_data)
 
 
 @pytest.mark.skipif(is_end_to_end(), reason="is not end-to-end test")
@@ -35,24 +34,22 @@ def test_use_provider_account_save(account_data, mock_tergiterc):
     for k, v in extras.items():
         expected_config += f"{k} = {v}\n"
 
-    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
-    provider = _tergite.use_provider_account(account, save=True)
+    provider = _tergite.use_provider_account(**account_data, save=True)
     with open(mock_tergiterc, "r") as file:
         actual_config = file.read()
 
     assert expected_config in actual_config
-    assert provider.account == account
+    assert provider.account == AccountInfo(**account_data)
 
 
 @pytest.mark.skipif(is_end_to_end(), reason="is not end-to-end test")
 @pytest.mark.parametrize("account_data", _PROVIDER_ACCOUNTS)
 def test_get_provider(account_data, mock_tergiterc):
     """get_provider returns the provider with the given service name"""
-    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
-    expected = _tergite.use_provider_account(account, save=True)
-    got = _tergite.get_provider(account.service_name)
+    expected = _tergite.use_provider_account(**account_data, save=True)
+    got = _tergite.get_provider(account_data["service_name"])
     assert got == expected
 
 
@@ -60,9 +57,8 @@ def test_get_provider(account_data, mock_tergiterc):
 @pytest.mark.parametrize("account_data", _PROVIDER_ACCOUNTS)
 def test_get_provider_non_existing(account_data, mock_tergiterc):
     """get_provider for non-existing service returns first provider"""
-    account = AccountInfo(**account_data)
     _tergite = factory.ProviderFactory(rc_file=mock_tergiterc)
-    _tergite.use_provider_account(account, save=True)
+    _tergite.use_provider_account(**account_data, save=True)
     provider = _tergite.get_provider(datetime.now().isoformat())
 
     with open(mock_tergiterc, "r") as file: